auth_service/requirements.txt
nameko==2.13.0
SQLAlchemy==1.4.25
nameko-sqlalchemy==1.5.0
pymysql==1.0.2
Werkzeug==2.0.1
PyJWT==2.3.0
cryptography==39.0.1
pytest>=7.0.0